Set default fzf command.
So I says to myself, I says - "Wow, this thing is really fast! It must be
using `rigrep` by default." WRONG. It was using *`find`* (!).

Unfortunately it seems like the default respect of .gitignore doesn't
check the global .gitignore rules, so had to pass some basics to the `glob`
flag.

https://github.com/junegunn/fzf.vim/issues/133
https://github.com/BurntSushi/ripgrep/issues/830
https://github.com/junegunn/fzf.vim/issues/133