Make a fuzzy cross-session tmux window jumper.
Inspired by [fzf wiki] examples. Instead of having one for panes, one
for windows, one for sessions, I wanted a single command to jump to
any window in any tmux session.

At first this was going to jump directly to a specific pane, but that
seemed too cumbersome to deal with picking from a list of 'bash' or
'vim' panes since panes don't have names in tmux. Also my workflow is
usually to have one 'action zone' per window, and switch to a pane with
Chris Toomey's marvelous [vim-tmux-navigator][2] is not so hard once
attached to the right window.

Also Moved some things around in .tmux.conf as it was getting fragmented.

[1]: https://github.com/junegunn/fzf/wiki/Examples#tmux
[2]: https://github.com/christoomey/vim-tmux-navigator

+shell +tmux +fzf