Trim fack file preview to small range of lines
Use `head` and `tail` to only show a small range of lines next to the
match. The match is the last line. Not great, but a slight improvement.

A better solution would be to use `sed -n 'NUM,NUM' FILE`, but the
nested command substitution and quotes is getting too gnarly to parse.

+shell +fzf +vim +ripgrep