Make linter ignores better
Update ALE's pattern options. Pass a setting to disable linting for any
buffer who's file name is '.min' or is in a folder called 'min'.

ALE uses Vim's [`match()`][1] against the file's full path. This is different
than a regular glob match. The existence of any match in the string
results in the setting being executed.

+vim +lint +ale

[1]: https://github.com/w0rp/ale/issues/371#issuecomment-304313091